requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-Mail==0.9.1
mysqlclient==2.2.4
bcrypt==4.1.3
python-dotenv==1.0.1
gunicorn==22.0.0
requests==2.32.3
